feat: refactor unused code
