+++ projects/portal/src/app/pages/nft-pawnshop/borrowers/borrow/borrow.component.html
