--- projects/portal/src/app/pages/nft-pawnshop/borrowers/borrower-nfts/borrower-nft/borrower-nft.component.html
@@ -1,14 +1,0 @@
-<view-borrower-nft
-  [classID]="classID$ | async"
-  [nftID]="nftID$ | async"
-  [listingInfo]="listingInfo$ | async"
-  [symbol]="symbol$ | async"
-  [symbolImage]="symbolImage$ | async"
-  [bids]="bids$ | async"
-  [loan]="loan$ | async"
-  [liquidation]="liquidation$ | async"
-  [nftMetadata]="nftMetadata$ | async"
-  [nftImage]="nftImage$ | async"
-  (appSubmitCancel)="onSubmitCancel($event)"
-  (appSubmitSell)="onSubmitSell($event)"
-></view-borrower-nft>
