+++ projects/portal/src/app/pages/nft-pawnshop/borrowers/list/list.component.html
