+++ projects/portal/src/app/pages/nft-pawnshop/borrowers/list/list.component.ts
@@ -39,7 +39,7 @@ export class ListComponent implements OnInit {
     );
   }
 
-  ngOnInit(): void { }
+  ngOnInit(): void {}
 
   onSubmit(data: ListRequest) {
     this.pawnshopApp.listNft(
