+++ projects/portal/src/app/pages/nft-pawnshop/borrowers/repay/repay.component.ts
@@ -1,42 +1,50 @@
 import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
+import cosmosclient from '@cosmos-client/core';
 import { BankQueryService } from 'projects/portal/src/app/models/cosmos/bank.query.service';
+import { BankService } from 'projects/portal/src/app/models/cosmos/bank.service';
 import { NftPawnshopApplicationService } from 'projects/portal/src/app/models/nft-pawnshops/nft-pawnshop.application.service';
-import { NftRequest } from 'projects/portal/src/app/models/nft-pawnshops/nft-pawnshop.model';
+import { NftPawnshopChartService } from 'projects/portal/src/app/models/nft-pawnshops/nft-pawnshop.chart.service';
+import { RepayRequest } from 'projects/portal/src/app/models/nft-pawnshops/nft-pawnshop.model';
 import { NftPawnshopQueryService } from 'projects/portal/src/app/models/nft-pawnshops/nft-pawnshop.query.service';
 import { NftPawnshopService } from 'projects/portal/src/app/models/nft-pawnshops/nft-pawnshop.service';
 import { Metadata } from 'projects/shared/src/lib/models/ununifi/query/nft/nft.model';
-import { Observable, combineLatest } from 'rxjs';
+import { Observable, combineLatest, of } from 'rxjs';
 import { map, mergeMap } from 'rxjs/operators';
 import {
-  ListedNfts200ResponseListingsInnerListing,
   BidderBids200ResponseBidsInner,
-  Loan200Response,
   Liquidation200ResponseLiquidations,
+  ListedNfts200ResponseListingsInnerListing,
 } from 'ununifi-client/esm/openapi';
 
 @Component({
-  selector: 'app-borrower-nft',
-  templateUrl: './borrower-nft.component.html',
-  styleUrls: ['./borrower-nft.component.css'],
+  selector: 'app-repay',
+  templateUrl: './repay.component.html',
+  styleUrls: ['./repay.component.css'],
 })
-export class BorrowerNftComponent implements OnInit {
+export class RepayComponent implements OnInit {
   classID$: Observable<string>;
   nftID$: Observable<string>;
   listingInfo$: Observable<ListedNfts200ResponseListingsInnerListing>;
   symbol$: Observable<string | null | undefined>;
+  symbolMetadataMap$: Observable<{
+    [symbol: string]: cosmosclient.proto.cosmos.bank.v1beta1.IMetadata;
+  }>;
   symbolImage$: Observable<string | undefined>;
   bids$: Observable<BidderBids200ResponseBidsInner[]>;
-  loan$: Observable<Loan200Response>;
   liquidation$: Observable<Liquidation200ResponseLiquidations>;
+  repayAmount$: Observable<number>;
   nftMetadata$: Observable<Metadata>;
   nftImage$: Observable<string>;
+  chartData$: Observable<(string | number)[][]>;
 
   constructor(
     private route: ActivatedRoute,
+    private readonly bankService: BankService,
     private readonly bankQuery: BankQueryService,
     private readonly pawnshop: NftPawnshopService,
     private readonly pawnshopQuery: NftPawnshopQueryService,
+    private readonly pawnshopChart: NftPawnshopChartService,
     private readonly pawnshopApp: NftPawnshopApplicationService,
   ) {
     this.classID$ = this.route.params.pipe(map((params) => params.class_id));
@@ -46,6 +54,7 @@ export class BorrowerNftComponent implements OnInit {
       mergeMap(([classID, nftID]) => this.pawnshopQuery.getNftListing$(classID, nftID)),
     );
     const denomMetadataMap$ = this.bankQuery.getDenomMetadataMap$();
+    this.symbolMetadataMap$ = this.bankQuery.getSymbolMetadataMap$();
     this.symbol$ = combineLatest([this.listingInfo$, denomMetadataMap$]).pipe(
       map(([info, metadata]) => metadata[info.bid_token || ''].symbol),
     );
@@ -54,16 +63,27 @@ export class BorrowerNftComponent implements OnInit {
     );
     this.bids$ = nftCombine$.pipe(
       mergeMap(([classID, nftID]) => this.pawnshopQuery.listNftBids$(classID, nftID)),
-      map((bids) =>
-        bids.sort((a, b) => parseInt(b.bid_amount?.amount!) - parseInt(a.bid_amount?.amount!)),
+      map((bidders) => bidders.filter((bidder) => bidder.borrowings?.length)),
+      map((bidders) =>
+        bidders.sort(
+          (a, b) => parseInt(b.deposit_amount?.amount!) - parseInt(a.deposit_amount?.amount!),
+        ),
       ),
-    );
-    this.loan$ = nftCombine$.pipe(
-      mergeMap(([classID, nftID]) => this.pawnshopQuery.getLoan$(classID, nftID)),
     );
     this.liquidation$ = nftCombine$.pipe(
       mergeMap(([classID, nftID]) => this.pawnshopQuery.getLiquidation$(classID, nftID)),
     );
+
+    this.repayAmount$ = combineLatest([this.liquidation$, denomMetadataMap$]).pipe(
+      map(
+        ([liquidation, denomMetadataMap]) =>
+          this.bankService.convertCoinToSymbolAmount(
+            liquidation.liquidation?.amount || { amount: '0', denom: '' },
+            denomMetadataMap,
+          ).amount,
+      ),
+    );
+
     const nftData$ = nftCombine$.pipe(
       mergeMap(([classID, nftID]) => this.pawnshopQuery.getNft$(classID, nftID)),
     );
@@ -73,15 +93,21 @@ export class BorrowerNftComponent implements OnInit {
     this.nftImage$ = nftData$.pipe(
       mergeMap((nft) => this.pawnshop.getImageFromUri(nft.nft?.uri || '')),
     );
+
+    this.chartData$ = this.bids$.pipe(
+      map((bids) => this.pawnshopChart.createBorrowingAmountChartData(bids)),
+      map((data) =>
+        data.sort(
+          (a, b) =>
+            Number((a[3] as string).replace('%', '')) - Number((b[3] as string).replace('%', '')),
+        ),
+      ),
+    );
   }
 
-  ngOnInit(): void { }
+  ngOnInit(): void {}
 
-  onSubmitCancel(data: NftRequest) {
-    this.pawnshopApp.cancelNftListing(data.classID, data.nftID);
-  }
-
-  onSubmitSell(data: NftRequest) {
-    this.pawnshopApp.sellingDecision(data.classID, data.nftID);
+  onSubmit(data: RepayRequest) {
+    this.pawnshopApp.repay(data.classID, data.nftID, data.repayBids);
   }
 }
