--- projects/portal/src/app/pages/nft-pawnshop/lenders/lender-nfts/lender-nft/lender-nft.component.html
@@ -1,15 +1,0 @@
-<view-lender-nft
-  [classID]="classID$ | async"
-  [nftID]="nftID$ | async"
-  [currentStoredWallet]="currentStoredWallet$ | async"
-  [listingInfo]="listingInfo$ | async"
-  [symbol]="symbol$ | async"
-  [symbolImage]="symbolImage$ | async"
-  [bidders]="bidders$ | async"
-  [nftMetadata]="nftMetadata$ | async"
-  [nftImage]="nftImage$ | async"
-  [bidDetail]="bidDetail$ | async"
-  [isWinning]="isWinning$ | async"
-  (appSubmitCancelBid)="onSubmitCancelBid($event)"
-  (appSubmitPayFullBid)="onSubmitPayFullBid($event)"
-></view-lender-nft>
