+++ projects/portal/src/app/pages/nft-pawnshop/lenders/place-bid/place-bid.component.html
