+++ projects/portal/src/app/pages/nft-pawnshop/nft-pawnshop.module.ts
@@ -1,27 +1,23 @@
-import { BorrowModule } from '../../views/nft-pawnshop/borrowers/borrower-nfts/borrower-nft/borrow/borrow.module';
-import { BorrowerNftModule } from '../../views/nft-pawnshop/borrowers/borrower-nfts/borrower-nft/borrower-nft.module';
-import { ListModule } from '../../views/nft-pawnshop/borrowers/borrower-nfts/borrower-nft/list/list.module';
-import { RepayModule } from '../../views/nft-pawnshop/borrowers/borrower-nfts/borrower-nft/repay/repay.module';
+import { BorrowModule } from '../../views/nft-pawnshop/borrowers/borrow/borrow.module';
 import { BorrowerModule } from '../../views/nft-pawnshop/borrowers/borrower/borrower.module';
 import { BorrowersModule } from '../../views/nft-pawnshop/borrowers/borrowers.module';
-import { LenderNftModule } from '../../views/nft-pawnshop/lenders/lender-nfts/lender-nft/lender-nft.module';
-import { PlaceBidModule } from '../../views/nft-pawnshop/lenders/lender-nfts/lender-nft/place-bid/place-bid.module';
+import { ListModule } from '../../views/nft-pawnshop/borrowers/list/list.module';
+import { RepayModule } from '../../views/nft-pawnshop/borrowers/repay/repay.module';
 import { LenderModule } from '../../views/nft-pawnshop/lenders/lender/lender.module';
 import { LendersModule } from '../../views/nft-pawnshop/lenders/lenders.module';
+import { PlaceBidModule } from '../../views/nft-pawnshop/lenders/place-bid/place-bid.module';
 import { NftPawnshopModule } from '../../views/nft-pawnshop/nft-pawnshop.module';
 import { ClassModule } from '../../views/nft-pawnshop/nfts/class/class.module';
 import { NftModule } from '../../views/nft-pawnshop/nfts/class/nft/nft.module';
 import { NftsModule } from '../../views/nft-pawnshop/nfts/nfts.module';
-import { BorrowComponent } from './borrowers/borrower-nfts/borrower-nft/borrow/borrow.component';
-import { BorrowerNftComponent } from './borrowers/borrower-nfts/borrower-nft/borrower-nft.component';
-import { ListComponent } from './borrowers/borrower-nfts/borrower-nft/list/list.component';
-import { RepayComponent } from './borrowers/borrower-nfts/borrower-nft/repay/repay.component';
+import { BorrowComponent } from './borrowers/borrow/borrow.component';
 import { BorrowerComponent } from './borrowers/borrower/borrower.component';
 import { BorrowersComponent } from './borrowers/borrowers.component';
-import { LenderNftComponent } from './lenders/lender-nfts/lender-nft/lender-nft.component';
-import { PlaceBidComponent } from './lenders/lender-nfts/lender-nft/place-bid/place-bid.component';
+import { ListComponent } from './borrowers/list/list.component';
+import { RepayComponent } from './borrowers/repay/repay.component';
 import { LenderComponent } from './lenders/lender/lender.component';
 import { LendersComponent } from './lenders/lenders.component';
+import { PlaceBidComponent } from './lenders/place-bid/place-bid.component';
 import { NftPawnshopRoutingModule } from './nft-pawnshop-routing.module';
 import { NftPawnshopComponent } from './nft-pawnshop.component';
 import { ClassComponent } from './nfts/class/class.component';
@@ -35,10 +31,8 @@ import { NgModule } from '@angular/core';
     NftPawnshopComponent,
     LendersComponent,
     LenderComponent,
-    LenderNftComponent,
     BorrowersComponent,
     BorrowerComponent,
-    BorrowerNftComponent,
     PlaceBidComponent,
     ListComponent,
     BorrowComponent,
@@ -53,10 +47,8 @@ import { NgModule } from '@angular/core';
     NftPawnshopModule,
     LendersModule,
     LenderModule,
-    LenderNftModule,
     BorrowersModule,
     BorrowerModule,
-    BorrowerNftModule,
     PlaceBidModule,
     ListModule,
     BorrowModule,
