+++ projects/portal/src/app/views/nft-pawnshop/borrowers/borrow/borrow.component.html
