+++ projects/portal/src/app/views/nft-pawnshop/borrowers/list/list.component.html
