--- projects/portal/src/app/views/nft-pawnshop/lenders/lender-nfts/lender-nft/lender-nft.component.html
@@ -1,165 +1,0 @@
-<div class="mx-auto max-w-screen-xl">
-  <div class="text-xl breadcrumbs">
-    <ul>
-      <li routerLink="../../../.."><a>Top</a></li>
-      <li routerLink="../../.."><a>Lending</a></li>
-      <li>{{ classID }}</li>
-      <li>{{ nftID }}</li>
-    </ul>
-  </div>
-  <div class="card lg:card-side bg-base-100 shadow-xl mb-8">
-    <figure class="aspect-square w-1/2 lg:w-1/4">
-      <!-- img width setting not working -->
-      <img
-        src="{{ nftImage || 'assets/UnUniFi-logo.png' }}"
-        alt="image of the NFT"
-        class="object-cover"
-      />
-    </figure>
-    <div class="card-body w-2/3">
-      <div class="flex flex-row">
-        <div class="badge badge-lg badge-primary">{{ listingInfo?.state || 'NOT_LISTING' }}</div>
-      </div>
-      <h2 class="card-title break-all">
-        {{ nftMetadata?.name || classID + '/' + nftID }}
-      </h2>
-      <div class="overflow-x-auto">
-        <table class="table w-full">
-          <tbody>
-            <tr>
-              <td>Class ID</td>
-              <td>{{ classID }}</td>
-            </tr>
-            <tr>
-              <td>NFT ID</td>
-              <td>{{ nftID }}</td>
-            </tr>
-            <tr>
-              <td>Owner</td>
-              <td>{{ listingInfo?.owner }}</td>
-            </tr>
-            <tr>
-              <td>Bid Token</td>
-              <td>{{ symbol }}</td>
-            </tr>
-            <tr>
-              <td>Minimum deposit rate</td>
-              <td>{{ listingInfo?.minimum_deposit_rate | percentage }}</td>
-            </tr>
-            <tr>
-              <td>Started at</td>
-              <td>{{ listingInfo?.started_at | date : 'yyyy-MM-dd a hh:mm:ss z' }}</td>
-            </tr>
-          </tbody>
-        </table>
-      </div>
-    </div>
-  </div>
-
-  <div class="card bg-base-100 shadow-xl mb-8">
-    <div class="card-body">
-      <div class="flex flex-row items-center">
-        <div class="stats stats-vertical md:stats-horizontal">
-          <div class="stat" *ngIf="bidders && bidders.length">
-            <div class="stat-title">Highest bid price</div>
-            <div class="stat-value text-primary">
-              {{ bidders[0].bid_amount?.amount | coinAmount }}
-            </div>
-            <div class="stat-desc">{{ bidders[0].bid_amount?.denom | coinDenom | async }}</div>
-          </div>
-          <div class="stat">
-            <div class="stat-title">Your Status</div>
-            <div *ngIf="!bidDetail" class="stat-value text-info">Not Offered</div>
-            <div *ngIf="bidDetail" class="stat-value text-accent">Offered</div>
-          </div>
-          <div
-            *ngIf="bidDetail && bidDetail.borrowings && bidDetail.borrowings.length"
-            class="stat"
-          >
-            <div class="stat-title">Lending Status</div>
-            <div class="stat-value text-primary">
-              {{ bidDetail.borrowings[0].amount | coin | async }}
-            </div>
-          </div>
-          <div *ngIf="isWinning" class="stat">
-            <div class="stat-title">Settlement Shortfall Deadline</div>
-            <div class="stat-value text-accent">
-              {{ listingInfo?.end_at | date : 'yyyy-MM-dd' }}
-            </div>
-            <div class="stat-desc">{{ listingInfo?.end_at | date : 'hh:mm:ss z' }}</div>
-          </div>
-        </div>
-        <span class="flex-auto"></span>
-        <div *ngIf="!bidDetail && !isWinning" class="card-actions justify-end">
-          <button routerLink="place-bid" class="btn btn-primary px-8">Make Offer</button>
-        </div>
-        <div
-          *ngIf="
-            bidDetail &&
-            !bidDetail.borrowings?.length &&
-            !isWinning &&
-            listingInfo?.state === 'BIDDING'
-          "
-          class="card-actions flex flex-row"
-        >
-          <button routerLink="place-bid" class="btn btn-primary px-8">Update Offer</button>
-          <span class="flex-auto"></span>
-          <button
-            *ngIf="bidders && bidders.length > 1"
-            class="btn btn-secondary px-8"
-            (click)="onSubmitCancelBid()"
-          >
-            Cancel Offer
-          </button>
-        </div>
-        <div *ngIf="isWinning" class="card-actions justify-end">
-          <button class="btn btn-primary px-8" (click)="onSubmitPayFullBid()">
-            Pay Settlement Shortfall
-          </button>
-        </div>
-      </div>
-    </div>
-  </div>
-
-  <div *ngIf="bidders?.length" class="card bg-base-100 shadow-xl">
-    <div class="overflow-x-auto">
-      <table class="table w-full">
-        <thead>
-          <tr>
-            <th></th>
-            <td>Interest</td>
-            <td>Deposit</td>
-            <td>Bid price</td>
-            <td>Expiry</td>
-            <td>Bidder</td>
-          </tr>
-        </thead>
-        <tbody>
-          <tr *ngFor="let bidder of bidders; let i = index">
-            <ng-container
-              *ngIf="bidder.id?.bidder === currentStoredWallet?.address; then color; else normal"
-            ></ng-container>
-            <ng-template #color>
-              <th class="text-secondary">#{{ i + 1 }}</th>
-              <td class="text-secondary">{{ bidder.deposit_lending_rate | percentage }}</td>
-              <td class="text-secondary">{{ bidder.deposit_amount | coin | async }}</td>
-              <td class="text-secondary">{{ bidder.bid_amount | coin | async }}</td>
-              <td class="text-secondary">
-                {{ bidder.bidding_period | date : 'yyyy-MM-dd a hh:mm z' }}
-              </td>
-              <td class="text-secondary">{{ bidder.id?.bidder }}</td>
-            </ng-template>
-            <ng-template #normal>
-              <th>#{{ i + 1 }}</th>
-              <td>{{ bidder.deposit_lending_rate | percentage }}</td>
-              <td>{{ bidder.deposit_amount | coin | async }}</td>
-              <td>{{ bidder.bid_amount | coin | async }}</td>
-              <td>{{ bidder.bidding_period | date : 'yyyy-MM-dd a hh:mm z' }}</td>
-              <td>{{ bidder.id?.bidder }}</td>
-            </ng-template>
-          </tr>
-        </tbody>
-      </table>
-    </div>
-  </div>
-</div>
