--- projects/portal/src/app/views/nft-pawnshop/lenders/lender-nfts/lender-nft/lender-nft.module.ts
@@ -1,14 +1,0 @@
-import { MaterialModule } from '../../../../material.module';
-import { LenderNftComponent } from './lender-nft.component';
-import { CommonModule } from '@angular/common';
-import { NgModule } from '@angular/core';
-import { FormsModule } from '@angular/forms';
-import { RouterModule } from '@angular/router';
-import { PipesModule } from 'projects/portal/src/app/pipes/pipes.module';
-
-@NgModule({
-  declarations: [LenderNftComponent],
-  imports: [CommonModule, RouterModule, FormsModule, MaterialModule, PipesModule],
-  exports: [LenderNftComponent],
-})
-export class LenderNftModule {}
