+++ projects/portal/src/app/views/nft-pawnshop/lenders/place-bid/place-bid.component.html
