+++ projects/portal/src/app/views/nft-pawnshop/nfts/class/nft/bidder-nft/bidder-nft.component.ts
@@ -21,7 +21,7 @@ export class BidderNftComponent implements OnInit {
   @Input()
   listingInfo?: ListedNfts200ResponseListingsInnerListing | null;
   @Input()
-  bidders?: BidderBids200ResponseBidsInner[] | null;
+  bids?: BidderBids200ResponseBidsInner[] | null;
   @Input()
   nftMetadata?: Metadata | null;
   @Input()
