+++ projects/portal/src/app/views/nft-pawnshop/nfts/class/nft/nft.component.html
@@ -9,7 +9,7 @@
       [classID]="classID"
       [nftID]="nftID"
       [listingInfo]="listingInfo"
-      [bidders]="bids"
+      [bids]="bids"
       [loan]="loan"
       [liquidation]="liquidation"
       [nftMetadata]="nftMetadata"
@@ -24,7 +24,7 @@
       [nftID]="nftID"
       [address]="address"
       [listingInfo]="listingInfo"
-      [bidders]="bids"
+      [bids]="bids"
       [nftMetadata]="nftMetadata"
       [nftImage]="nftImage"
       [ownBid]="ownBid"
